Make the Reporting warehouse functional: schema provisioning + the missing projections (REP-SCHEMA)
The reporting warehouse had no schema provisioning at all — empty migrations, no startup
initializer, no creation script — so the projection consumers would have thrown "table
doesn't exist" on the first live event and the service was non-functional (which is why
the customer-journey E2E never exercised it). Three analytical tables read by the
dashboards (inventory_snapshot, customer_segment_snapshot, delivery_facts) additionally
had no writer at all.

Provisioning: a ReportingWarehouseInitializer hosted service now creates the schema on
startup — EF Core EnsureCreated for the invoice and inbox tables it owns (it also creates
the database on a fresh server), then the new Dapper ReportingWarehouseSchema for the
analytical fact and snapshot tables. Both steps are idempotent. The AppHost creates the
reportingdb MySQL database so the initializer has something to connect to. The test
fixture now provisions from the same ReportingWarehouseSchema, so tests exercise the exact
production DDL.

Missing writers: the projection writer now feeds all three previously-unwritten tables.
ProductCreated seeds inventory_snapshot; DeliveryScheduled records the slot in
delivery_facts and DeliveryCompleted stamps it with the derived on-time/late outcome and
duration (throwing to redeliver if the scheduled row has not arrived yet); and the
OrderConfirmed projection now also upserts customer_segment_snapshot, segmenting the
customer as new on their first order and returning thereafter — all inside the existing
exactly-once inbox transaction. Three consumers wire the events in.

Also fixed a latent bug the missing data had hidden: the inventory, delivery and customer
dashboard reads materialised MySQL COUNT (BIGINT) and SUM (DECIMAL) aggregates into int
record constructors, which Dapper cannot bind — the dashboards would have 500'd. The
summary counts are now long and the conditional counts use COUNT(CASE ...) so the types
line up.

Proven against a real MySQL warehouse: the schema provisions every analytical table; a
ProductCreated seeds an inventory snapshot exactly once; a scheduled-then-completed
delivery is recorded on time and a late completion as late; completing before the schedule
arrives throws for retry; and a customer's first order segments them as new and a second
as returning, read back through the real dashboard queries. Reporting suite 54 green; full
solution builds 0/0.

+++ src/Services/Reporting/FreshCart.Reporting.Application/Common/Abstractions/CustomerAcquisitionSummary.cs
@@ -4,7 +4,7 @@ namespace FreshCart.Reporting.Application.Common.Abstractions;
 /// Acquisition and retention headline numbers for a reporting period.
 /// </summary>
 public sealed record CustomerAcquisitionSummary(
-    int NewCustomers,
-    int ReturningCustomers,
-    int ChurnedCustomers,
+    long NewCustomers,
+    long ReturningCustomers,
+    long ChurnedCustomers,
     decimal AverageLifetimeValue);
